requirements.txt
azure-storage-blob
cryptography
httpx